+++ bld/wres/h/layer2.h
@@ -31,8 +31,6 @@
 
 
 extern bool InitResources2( struct WResDirHead **, PHANDLE_INFO hinfo );
-extern int  WResLoadString2( struct WResDirHead *, PHANDLE_INFO hinfo,
-                            UINT idResource, LPSTR lpszBuffer, int nBufferMax );
-extern int  WResLoadResource2( struct WResDirHead *, PHANDLE_INFO hinfo, WResID *idType,
-                              WResID *idResource, LPSTR *lpszBuffer, int *bufferSize );
+extern int  WResLoadString2( struct WResDirHead *, PHANDLE_INFO hinfo, UINT idResource, LPSTR lpszBuffer, int nBufferMax );
+extern int  WResLoadResource2( struct WResDirHead *, PHANDLE_INFO hinfo, WResID *idType, WResID *idResource, LPSTR *lpszBuffer, size_t *bufferSize );
 extern bool FiniResources2( struct WResDirHead *, PHANDLE_INFO hinfo );
